refactor(search): unifica estado dos painéis em openPanel e decompõe SearchForm
SearchForm.tsx tinha 1005 linhas e o orquestrador carregava 5 booleans
paralelos para a abertura dos painéis (dest, calendar, guests, trip,
budget), que são mutuamente exclusivos.

- Substitui os 5 `show*` por uma união discriminada `openPanel`:
  toggles viram `setOpenPanel(p => p === x ? null : x)`, `closeAllPanels`
  vira `setOpenPanel(null)`. O type `CloseHandlers` e metade dos
  useCallback de fechamento somem.
- `useSearchFormDismiss` agora itera sobre um registry `{ panel, ref }[]`
  em vez de 5 branches idênticas de click-outside.
- Move os sub-campos memoizados (DestinationField, DateField,
  GuestsField, TripTypeField, BudgetField, SearchButton) + helpers +
  tipos + hook para `components/search/`. SearchForm.tsx cai para 374
  linhas (orquestrador + estado).
- Corrige warning de chave duplicada no header de dias da semana do
  calendário (Qua/Qui, Seg/Sex/Sab colidiam).

Comportamento de UI (abrir/fechar, click-outside, Esc) inalterado,
coberto por tests/e2e/hero_ux.spec.ts.

Closes #967
